Move CompatibilitySpec to primitive
